% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_strata_ads.R
\name{batch_strata_ads}
\alias{batch_strata_ads}
\title{Create Facebook ads for a batch of strata}
\usage{
batch_strata_ads(fbacc, strata_targeting, optimization_goal = c("NONE",
  "APP_INSTALLS", "CLICKS", "ENGAGED_USERS", "EXTERNAL", "EVENT_RESPONSES",
  "IMPRESSIONS", "LINK_CLICKS", "OFFER_CLAIMS", "OFFSITE_CONVERSIONS",
  "PAGE_ENGAGEMENT", "PAGE_LIKES", "POST_ENGAGEMENT", "REACH",
  "SOCIAL_IMPRESSIONS", "VIDEO_VIEWS"), billing_event = c("APP_INSTALLS",
  "CLICKS", "IMPRESSIONS", "LINK_CLICKS", "OFFER_CLAIMS", "PAGE_LIKES",
  "POST_ENGAGEMENT", "VIDEO_VIEWS"), promoted_object, campaign_id,
  wait_interval = 100, show_wait, ...)
}
\arguments{
\item{fbacc}{\code{FB_Ad_account} object that identifies your Facebook Ad account. Note this is not your Ad account number.}

\item{strata_targeting}{list. Information about each strata that you plan to target.}

\item{optimization_goal}{character. Optimization goal.}

\item{billing_event}{character. The billing event.}

\item{promoted_object}{list. See at \url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign/promoted-object}.}

\item{campaign_id}{character. Parent Ad Campaign ID.}

\item{wait_interval}{integer. Number of seconds to wait between creation of each ad to avoid exceeding the API rate limit.}

\item{show_wait}{logical. If \code{TRUE}, the console will display a wait time progress bar.}

\item{...}{further arguments passed to the API endpoint}
}
\value{
A data.frame with Strata IDs, Ad Set IDs, Ad IDs for ads that have been successfully created.
}
\description{
Create Facebook ads for a batch of strata
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign#Creating}
}
