% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target.R
\name{create_target}
\alias{create_target}
\title{Create targeting attributes for a strata}
\usage{
create_target(strata_id, geo_locations, age_min = NULL, age_max = NULL,
  genders = NULL, education_statuses = NULL, ethnic_affinity = NULL,
  interests = NULL, behaviors = NULL, exclusions = NULL,
  other_demographics = NULL, is_autobid = FALSE, bid_amount = NULL,
  daily_budget, adset_status = c("ACTIVE", "PAUSED"),
  ad_status = c("ACTIVE", "PAUSED"), start_time = NULL, end_time = NULL,
  creative_id)
}
\arguments{
\item{strata_id}{character. Strata ID.}

\item{geo_locations}{list. A list specifying geographic location of the strata. See \url{https://developers.facebook.com/docs/marketing-api/buying-api/targeting#location}.}

\item{age_min}{integer.     Minimum age. If used, must be 13 or higher. Default to 18.}

\item{age_max}{integer. Maximum age. If used, must be 65 or lower.}

\item{genders}{integer. 1=male, 2=female.}

\item{education_statuses}{integer. Vector of Facebook IDs to target based on education level. See \url{https://developers.facebook.com/docs/marketing-api/targeting-specs/}.}

\item{ethnic_affinity}{integer. Vector of Facebook IDs for each ethnic group.}

\item{interests}{list. A list of Facebook IDs for interests.}

\item{behaviors}{list. A list of Facebook IDs for behaviors.}

\item{exclusions}{list. A list of Facebook IDs for excluding certain demographics.}

\item{other_demographics}{list. A list of Facebook IDs for other demographic targeting parameters.}

\item{is_autobid}{logical. If \code{TRUE}, bidding amount is automated and you do not need to include a \code{bid_amount}.}

\item{bid_amount}{integer. The amount you set for bid and budget are at ad account currencies minimum denomination level. For example cents for US dollars.}

\item{daily_budget}{integer. The amount you set for bid and budget are at ad account currencies minimum denomination level. For example cents for US dollars.}

\item{adset_status}{character. Ad Set status.}

\item{ad_status}{character. Ad status.}

\item{start_time}{(optinal) UTC UNIX timestamp.}

\item{end_time}{(optional) UTC UNIX timestamp.}

\item{creative_id}{character. Creative ID that identifies the ad creative object you plan to display as your ad. See \url{https://developers.facebook.com/docs/marketing-api/reference/ad-creative}.}
}
\value{
A list with targeting attributes for a strata that can be used to create ads.
}
\description{
Create targeting attributes for a strata
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/buying-api/targeting}
}
