% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb_edu_level}
\alias{fb_edu_level}
\title{Education Level}
\format{A dataset with two variables: \code{key} and \code{description}.}
\usage{
fb_edu_level
}
\description{
A dataset that contains the keys for targeting specific levels of education.
}
\keyword{datasets}
