% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb_ethnicity}
\alias{fb_ethnicity}
\title{U.S. Ethnic Affinity}
\format{A data frame with six variables including the ethnic affinity \code{id} and
\code{description}, along with metadata from the API Targeting Search results.}
\usage{
fb_ethnicity
}
\description{
A dataset that contains the ids for targeting specific ethnic groups in the U.S.
}
\keyword{datasets}
