% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb_politics}
\alias{fb_politics}
\title{U.S. Politics}
\format{A dataset with 10 variables including the political
ideologies/behaviors \code{id} and \code{description}, along with metadata
from the API Targeting Search results.}
\usage{
data(fb_politics)
}
\description{
A data set that contains the ids for targeting specific political ideologies
or behaviors in the U.S.
}
\keyword{datasets}
