% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb_US_states}
\alias{fb_US_states}
\title{U.S. States and Regions}
\format{A dataset with 5 variables including the \code{key} for each of the 50 U.S.
states along with its corresponding Census region.}
\usage{
fb_US_states
}
\description{
A data set that contains the keys for targeting states or regions within the U.S.
}
\keyword{datasets}
