% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_ad.R
\name{strata_ad}
\alias{strata_ad}
\title{Create a Facebook ad for a strata}
\usage{
strata_ad(fbacc, strata_id, adset_name, ad_name, optimization_goal = c("NONE",
  "APP_INSTALLS", "CLICKS", "ENGAGED_USERS", "EXTERNAL", "EVENT_RESPONSES",
  "IMPRESSIONS", "LINK_CLICKS", "OFFER_CLAIMS", "OFFSITE_CONVERSIONS",
  "PAGE_ENGAGEMENT", "PAGE_LIKES", "POST_ENGAGEMENT", "REACH",
  "SOCIAL_IMPRESSIONS", "VIDEO_VIEWS"), billing_event = c("APP_INSTALLS",
  "CLICKS", "IMPRESSIONS", "LINK_CLICKS", "OFFER_CLAIMS", "PAGE_LIKES",
  "POST_ENGAGEMENT", "VIDEO_VIEWS"), bid_amount, promoted_object, campaign_id,
  is_autobid = FALSE, status = c("ACTIVE", "PAUSED", "ARCHIVED", "DELETED"),
  daily_budget, end_time = NULL, start_time = NULL, creative_id, targeting,
  adset_status = c("ACTIVE", "PAUSED"), ad_status = c("ACTIVE", "PAUSED"),
  ...)
}
\arguments{
\item{fbacc}{\code{FB_Ad_account} object that identifies your Facebook Ad account. Note this is not your Ad account number.}

\item{strata_id}{character. Strata ID.}

\item{adset_name}{character. Name of the Ad Set.}

\item{optimization_goal}{character. Optimization goal.}

\item{billing_event}{character. The billing event.}

\item{bid_amount}{integer. The amount you set for bid and budget are at ad account currencies minimum denomination level. For example cents for US dollars.}

\item{promoted_object}{list. See at \url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign/promoted-object}.}

\item{campaign_id}{character. Parent Ad Campaign ID.}

\item{is_autobid}{logical. If \code{TRUE}, bidding amount is automated and you do not need to include a \code{bid_amount}.}

\item{daily_budget}{integer. The amount you set for bid and budget are at ad account currencies minimum denomination level. For example cents for US dollars.}

\item{end_time}{UTC UNIX timestamp.}

\item{start_time}{UTC UNIX timestamp.}

\item{creative_id}{character. Creative ID that identifies the ad creative object you plan to display as your ad. See \url{https://developers.facebook.com/docs/marketing-api/reference/ad-creative}.}

\item{targeting}{list. Describes the demographics of your strata.}

\item{adset_status}{character. Ad Set status.}

\item{ad_status}{character. Ad status.}

\item{...}{further arguments passed to the API endpoint}
}
\value{
A data.frame with Strata ID, Ad Set ID, Ad ID
}
\description{
Create a Facebook ad for a strata
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign#Creating}
}
